+++ CHANGELOG.md
@@ -19,6 +19,8 @@ This means you can reuse the same neural networks between for example 0.2.0 and 
 
 - `changed` Default neural nets are now compiled into the executable.
 - `changed` Rollouts and the training process are now deterministic.
+- `changed` Reverted from `Hardsigmoid` back to `ReLU`.
+- `changed` Use `CrossEntropyLoss` as PyTorch Optimizer and add `softmax` only after training.
 
 ## 0.2.0 - 2023-11-26